+++ src/Service/ReqserLanguageRedirectService.php
@@ -361,7 +361,6 @@ class ReqserLanguageRedirectService
 
         // Handle alternative language redirect if configured
         if (($this->redirectConfig['redirectToAlternativeLanguage'] ?? false) &&
-            isset($this->primaryBrowserLanguage) && //Added from 2.0.5 to prevent bots to be redirected
             !in_array($this->primaryBrowserLanguage, $this->getAlternativeBrowserLanguages()) &&
             isset($this->domainMappings['domainLanguageCode'][$this->redirectConfig['alternativeRedirectLanguageCode']])) {